backend/requirements.txt
fastapi==0.70.0
fastapi-mail==1.0.2
fastapi-pagination==0.9.1
alembic==1.7.5
Authlib==0.15.5
email-validator==1.1.3
Faker==10.0.0
celery==5.2.1
httpx==0.19.0
ipython==7.30.1
itsdangerous==1.1.0
pytest==6.2.5
requests==2.28.1
SQLAlchemy==1.4.28
SQLAlchemy-Utils==0.37.9
uvicorn==0.16.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.5
PyJWT==1.7.1
#psycopg==3.0.7
psycopg2==2.9.1
fastapi-mail==1.0.2
fastapi-pagination==0.9.1
Jinja2==3.0.3
secure==0.3.0
mysqlclient==2.1.1
eth-brownie==1.19.1
APScheduler==3.9.1